+++ dev-docs/engineering/event-catalog.md
@@ -6,7 +6,7 @@
 | --- | --- |
 | `submission` | noj-core/src/mq/consumer.ts<br>noj-core/src/routes/sse.ts |
 | `queue` | noj-core/src/domains/submission/services/queue.ts<br>noj-core/src/domains/submission/services/self-tests.ts<br>noj-core/src/domains/submission/services/submissions/artifact-submissions.ts<br>noj-core/src/domains/submission/services/submissions/submissions-crud.ts<br>noj-core/src/domains/submission/services/submissions/submissions-rejudge.ts<br>noj-core/src/mq/consumer.ts<br>noj-core/src/routes/sse.ts |
-| `user` | noj-core/src/domains/messaging/routes/conversations.ts<br>noj-core/src/domains/messaging/services/messages.ts<br>noj-core/src/routes/sse.ts<br>noj-core/src/services/notifications.ts |
+| `user` | noj-core/src/domains/community/services/notifications.ts<br>noj-core/src/domains/messaging/routes/conversations.ts<br>noj-core/src/domains/messaging/services/messages.ts<br>noj-core/src/routes/sse.ts |
 | `contestRanking` | noj-core/src/domains/submission/services/submissions/submissions-result.ts<br>noj-core/src/routes/sse.ts |
 | `contestSubmission` | noj-core/src/domains/submission/services/submissions/artifact-submissions.ts<br>noj-core/src/domains/submission/services/submissions/submissions-crud.ts<br>noj-core/src/routes/sse.ts |
 | `stats` | noj-core/src/routes/sse.ts<br>noj-core/src/services/stats-cache.ts |
