+++ dev-docs/engineering/route-catalog.md
@@ -5,55 +5,23 @@
 | 方法 | 路径 | 文件 |
 | --- | --- | --- |
 | DELETE | `/:id` | noj-core/src/routes/admin-announcements.ts |
-| DELETE | `/admin/boards/:boardId/role-grants/:roleId` | noj-core/src/routes/community-admin.ts |
-| DELETE | `/admin/sanctions/:sanctionId` | noj-core/src/routes/community-admin.ts |
-| DELETE | `/comments/:commentId` | noj-core/src/routes/community.ts |
-| DELETE | `/posts/:postId` | noj-core/src/routes/community.ts |
 | GET | `/` | noj-core/src/routes/admin-announcements.ts |
 | GET | `/` | noj-core/src/routes/announcements.ts |
 | GET | `/` | noj-core/src/routes/rankings.ts |
 | GET | `/` | noj-core/src/routes/search.ts |
 | GET | `/:id` | noj-core/src/routes/announcements.ts |
-| GET | `/admin/boards/:boardId/role-grants` | noj-core/src/routes/community-admin.ts |
-| GET | `/admin/comments/pending` | noj-core/src/routes/community-admin.ts |
-| GET | `/admin/reports` | noj-core/src/routes/community-admin.ts |
-| GET | `/admin/sanctions` | noj-core/src/routes/community-admin.ts |
-| GET | `/admin/users/:userId/sanctions` | noj-core/src/routes/community-admin.ts |
-| GET | `/boards` | noj-core/src/routes/community.ts |
-| GET | `/bookmarks` | noj-core/src/routes/community.ts |
 | GET | `/checkin` | noj-core/src/routes/rankings.ts |
 | GET | `/community/notifications/events` | noj-core/src/routes/sse.ts |
-| GET | `/config` | noj-core/src/routes/community.ts |
 | GET | `/contests/:id/events` | noj-core/src/routes/sse.ts |
 | GET | `/events` | noj-core/src/routes/announcements.ts |
-| GET | `/feed` | noj-core/src/routes/community.ts |
 | GET | `/health` | noj-core/src/routes/health.ts |
 | GET | `/me` | noj-core/src/routes/rankings.ts |
-| GET | `/notifications` | noj-core/src/routes/community.ts |
-| GET | `/notifications/unread-count` | noj-core/src/routes/community.ts |
-| GET | `/posts` | noj-core/src/routes/community.ts |
-| GET | `/posts/:postId` | noj-core/src/routes/community.ts |
-| GET | `/posts/:postId/comments` | noj-core/src/routes/community.ts |
-| GET | `/posts/counts` | noj-core/src/routes/community.ts |
 | GET | `/queue/events` | noj-core/src/routes/sse.ts |
-| GET | `/reports/:reportId` | noj-core/src/routes/community.ts |
-| GET | `/solutions/eligibility` | noj-core/src/routes/community.ts |
 | GET | `/stats` | noj-core/src/routes/stats.ts |
 | GET | `/submissions/:id/events` | noj-core/src/routes/sse.ts |
 | GET | `/submissions/stats/events` | noj-core/src/routes/sse.ts |
 | GET | `NOJ_ENV` | noj-core/src/routes/health.ts |
 | GET | `userId` | noj-core/src/routes/admin-announcements.ts |
-| GET | `userId` | noj-core/src/routes/community-admin.ts |
-| GET | `userId` | noj-core/src/routes/community.ts |
-| GET | `userId` | noj-core/src/routes/community.ts |
-| GET | `userId` | noj-core/src/routes/community.ts |
-| GET | `userId` | noj-core/src/routes/community.ts |
-| GET | `userId` | noj-core/src/routes/community.ts |
-| GET | `userId` | noj-core/src/routes/community.ts |
-| GET | `userId` | noj-core/src/routes/community.ts |
-| GET | `userId` | noj-core/src/routes/community.ts |
-| GET | `userId` | noj-core/src/routes/community.ts |
-| GET | `userId` | noj-core/src/routes/community.ts |
 | GET | `userId` | noj-core/src/routes/search.ts |
 | GET | `userId` | noj-core/src/routes/sse.ts |
 | GET | `userId` | noj-core/src/routes/sse.ts |
@@ -62,27 +30,5 @@
 | GET | `userRole` | noj-core/src/routes/search.ts |
 | GET | `userRole` | noj-core/src/routes/sse.ts |
 | GET | `userRole` | noj-core/src/routes/sse.ts |
-| PATCH | `/admin/boards/:boardId` | noj-core/src/routes/community-admin.ts |
-| PATCH | `/comments/:commentId` | noj-core/src/routes/community.ts |
-| PATCH | `/posts/:postId` | noj-core/src/routes/community.ts |
 | POST | `/` | noj-core/src/routes/admin-announcements.ts |
-| POST | `/admin/boards` | noj-core/src/routes/community-admin.ts |
-| POST | `/admin/comments/:commentId/:status` | noj-core/src/routes/community-admin.ts |
-| POST | `/admin/posts/:postId/:flag` | noj-core/src/routes/community-admin.ts |
-| POST | `/admin/posts/:postId/:status` | noj-core/src/routes/community-admin.ts |
-| POST | `/admin/preset/:preset` | noj-core/src/routes/community-admin.ts |
-| POST | `/admin/reports/:reportId/:status` | noj-core/src/routes/community-admin.ts |
-| POST | `/admin/reports/:reportId/reopen` | noj-core/src/routes/community-admin.ts |
-| POST | `/admin/sanctions` | noj-core/src/routes/community-admin.ts |
-| POST | `/comments/:commentId/like` | noj-core/src/routes/community.ts |
-| POST | `/notifications/:id/read` | noj-core/src/routes/community.ts |
-| POST | `/notifications/read` | noj-core/src/routes/community.ts |
-| POST | `/posts` | noj-core/src/routes/community.ts |
-| POST | `/posts/:postId/bookmark` | noj-core/src/routes/community.ts |
-| POST | `/posts/:postId/comments` | noj-core/src/routes/community.ts |
-| POST | `/posts/:postId/like` | noj-core/src/routes/community.ts |
-| POST | `/reports` | noj-core/src/routes/community.ts |
-| POST | `/users/:userId/follow` | noj-core/src/routes/community.ts |
 | PUT | `/:id` | noj-core/src/routes/admin-announcements.ts |
-| PUT | `/admin/boards/:boardId/role-grants/:roleId` | noj-core/src/routes/community-admin.ts |
-| PUT | `/me/activity-visibility` | noj-core/src/routes/community.ts |
