+++ noj-core/src/app.ts
@@ -16,8 +16,8 @@ import selfTests from "./domains/submission/routes/self-tests.ts";
 import users from "./domains/identity/routes/users.ts";
 import rankings from "./routes/rankings.ts";
 import conversations from "./domains/messaging/routes/conversations.ts";
-import community from "./routes/community.ts";
-import communityAdmin from "./routes/community-admin.ts";
+import community from "./domains/community/routes/community.ts";
+import communityAdmin from "./domains/community/routes/community-admin.ts";
 import search from "./routes/search.ts";
 import contests from "./domains/contest/routes/contests.ts";
 import trainings from "./domains/catalog/routes/trainings.ts";
