+++ noj-core/src/domains/community/services/community/community-boards.ts
@@ -1,8 +1,11 @@
 import { and, eq } from "drizzle-orm";
-import { getDb } from "../../db/connection.ts";
-import { communityBoardRoleGrants, communityBoards } from "../../db/schema.ts";
-import { NotFoundError } from "../../lib/errors.ts";
-import { nowIso } from "../../lib/dates.ts";
+import { getDb } from "../../../../db/connection.ts";
+import {
+  communityBoardRoleGrants,
+  communityBoards,
+} from "../../../../db/schema.ts";
+import { NotFoundError } from "../../../../lib/errors.ts";
+import { nowIso } from "../../../../lib/dates.ts";
 
 export function listBoards(includeArchived = false) {
   const db = getDb();
