+++ noj-core/src/domains/community/services/community/community-comments.ts
@@ -1,17 +1,17 @@
 import { and, eq, ne, or, sql } from "drizzle-orm";
-import { getDb } from "../../db/connection.ts";
+import { getDb } from "../../../../db/connection.ts";
 import {
   communityComments,
   communityModerationActions,
   communityPosts,
   users,
-} from "../../db/schema.ts";
+} from "../../../../db/schema.ts";
 import {
   ForbiddenError,
   NotFoundError,
   ValidationError,
-} from "../../lib/errors.ts";
-import { logAudit } from "../audit-log.ts";
+} from "../../../../lib/errors.ts";
+import { logAudit } from "../../../system/index.ts";
 import { createNotification } from "../notifications.ts";
 import {
   assertCommunityEnabled,
@@ -19,7 +19,7 @@ import {
 } from "./community-config.ts";
 import { getPost } from "./community-post-crud.ts";
 import { publicationStatus } from "./community-post-common.ts";
-import { nowIso } from "../../lib/dates.ts";
+import { nowIso } from "../../../../lib/dates.ts";
 
 /** 审核/处置评论状态：批准待审评论时补发回复通知（原 pending 创建时不发）。 */
 export async function changeCommentStatus(
