+++ noj-core/src/domains/community/services/community/community-post-list.ts
@@ -1,7 +1,11 @@
 import { and, desc, eq, ilike, inArray, lt, ne, or, sql } from "drizzle-orm";
-import { getDb } from "../../db/connection.ts";
-import { communityBookmarks, communityPosts, users } from "../../db/schema.ts";
-import type { CommunityPostType } from "../../types/community.ts";
+import { getDb } from "../../../../db/connection.ts";
+import {
+  communityBookmarks,
+  communityPosts,
+  users,
+} from "../../../../db/schema.ts";
+import type { CommunityPostType } from "../../../../types/community.ts";
 import {
   assertCommunityEnabled,
   getCommunityConfig,
