+++ noj-core/src/services/seed/seed-rbac.ts
@@ -22,7 +22,7 @@ import {
 } from "../../db/schema.ts";
 import { PERMISSION_DEFS } from "../../types/index.ts";
 import { ROOT_USER_ID } from "../../lib/constants.ts";
-import { ensureCommunitySeeds } from "../community/community-seed.ts";
+import { ensureCommunitySeeds } from "../../domains/community/index.ts";
 
 // user 角色的默认权限（action 列表）
 const USER_DEFAULT_PERMISSIONS: Array<{ resource: string; action: string }> = [
